requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
psycopg2-binary
asyncpg
sqlalchemy

